src/trading/executor.py
from typing import Dict, Any
import os
from dotenv import load_dotenv

class TradeExecutor:
    def __init__(self):
        load_dotenv()
        self.paper_trading = os.getenv("ENABLE_PAPER_TRADING", "true").lower() == "true"

    def execute(self, symbol: str, side: str, quantity: float) -> Dict[str, Any]:
        """
        Execute a trade order
        """
        if self.paper_trading:
            return self._paper_trade(symbol, side, quantity)
        else:
            return self._live_trade(symbol, side, quantity)

    def _paper_trade(self, symbol: str, side: str, quantity: float) -> Dict[str, Any]:
        """
        Simulate trade execution for paper trading
        """
        return {
            "status": "success",
            "type": "paper",
            "symbol": symbol,
            "side": side,
            "quantity": quantity,
            "timestamp": "2025-08-10T12:00:00Z"  # Replace with actual timestamp
        }

    def _live_trade(self, symbol: str, side: str, quantity: float) -> Dict[str, Any]:
        """
        Execute live trade through broker/exchange API
        """
        # Implement your broker/exchange API integration here
        raise NotImplementedError("Live trading not implemented yet")
