requirements.txt
python-dotenv==1.0.0
pandas==2.1.0
numpy==1.24.3
requests==2.31.0
yfinance==0.2.28
python-binance==1.0.19
pytest==7.4.0
fastapi==0.103.0
uvicorn==0.23.2
